+++ api/src/org/labkey/api/exp/api/ExpProtocolApplication.java
@@ -63,7 +63,7 @@ public interface ExpProtocolApplication extends ExpObject
     @NotNull ExpDataRunInput addDataInput(User user, ExpData input, String inputRole, @Nullable ExpDataProtocolInput protocolInput);
     void removeAllDataInputs(User user);
     void removeDataInput(User user, ExpData data);
-    void removeDataInputs(User user, Collection<Integer> rowIds);
+    void removeDataInputs(User user, Collection<Long> rowIds);
 
     /**
      * Add a material input
@@ -72,10 +72,10 @@ public interface ExpProtocolApplication extends ExpObject
      */
     @NotNull ExpMaterialRunInput addMaterialInput(User user, ExpMaterial material, @Nullable String inputRole);
     @NotNull ExpMaterialRunInput addMaterialInput(User user, ExpMaterial material, @Nullable String inputRole, @Nullable ExpMaterialProtocolInput protocolInput);
-    void addMaterialInputs(User user, Collection<Integer> materialRowIds, @Nullable String inputRole, @Nullable ExpMaterialProtocolInput protocolInput);
+    void addMaterialInputs(User user, Collection<Long> materialRowIds, @Nullable String inputRole, @Nullable ExpMaterialProtocolInput protocolInput);
     void removeAllMaterialInputs(User user);
     void removeMaterialInput(User user, ExpMaterial material);
-    void removeMaterialInputs(User user, Collection<Integer> rowIds);
+    void removeMaterialInputs(User user, Collection<Long> rowIds);
 
     void addProvenanceInput(Set<String> lsids);
     void addProvenanceMapping(Set<Pair<String, String>> lsidPairs);
